# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Near-Earth Objects Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CloseApproach {
  date: string;
  missDistanceKm: number;
  missDistanceLunar: number;
  missDistanceAu: number;
  relativeVelocityKmS: number;
  relativeVelocityKmH: number;
  orbitingBody: string;
}

export interface NeoObject {
  id: string;
  name: string;
  nasaJplUrl: string;
  absoluteMagnitudeH: number;
  estimatedDiameterMinKm: number;
  estimatedDiameterMaxKm: number;
  isPotentiallyHazardous: boolean;
  isSentryObject: boolean;
  closeApproaches: CloseApproach[];
  nextCloseApproach: CloseApproach;
  orbitClass: string;
  /** @nullable */
  orbitClassDescription: string | null;
}

export interface NeoListResponse {
  objects: NeoObject[];
  fetchedAt: string;
  dateRangeStart: string;
  dateRangeEnd: string;
}

export interface NeoStats {
  totalCount: number;
  hazardousCount: number;
  closestApproachKm: number;
  closestObjectName: string;
  fastestVelocityKmS: number;
  fastestObjectName: string;
  largestDiameterKm: number;
  largestObjectName: string;
  sentryCount: number;
  fetchedAt: string;
}

export type GetApproachingNeoParams = {
/**
 * Number of days to look ahead (max 7 per NASA API)
 * @minimum 1
 * @maximum 7
 */
days?: number;
};