# lib/api-zod/src/generated/types/closeApproach.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Near-Earth Objects Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CloseApproach {
  date: string;
  missDistanceKm: number;
  missDistanceLunar: number;
  missDistanceAu: number;
  relativeVelocityKmS: number;
  relativeVelocityKmH: number;
  orbitingBody: string;
}