# lib/api-zod/src/generated/types/neoListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Near-Earth Objects Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { NeoObject } from './neoObject';

export interface NeoListResponse {
  objects: NeoObject[];
  fetchedAt: string;
  dateRangeStart: string;
  dateRangeEnd: string;
}