# lib/api-zod/src/generated/types/neoObject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Near-Earth Objects Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CloseApproach } from './closeApproach';

export interface NeoObject {
  id: string;
  name: string;
  nasaJplUrl: string;
  absoluteMagnitudeH: number;
  estimatedDiameterMinKm: number;
  estimatedDiameterMaxKm: number;
  isPotentiallyHazardous: boolean;
  isSentryObject: boolean;
  closeApproaches: CloseApproach[];
  nextCloseApproach: CloseApproach;
  orbitClass: string;
  /** @nullable */
  orbitClassDescription: string | null;
}