# lib/api-zod/src/generated/types/neoStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Near-Earth Objects Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface NeoStats {
  totalCount: number;
  hazardousCount: number;
  closestApproachKm: number;
  closestObjectName: string;
  fastestVelocityKmS: number;
  fastestObjectName: string;
  largestDiameterKm: number;
  largestObjectName: string;
  sentryCount: number;
  fetchedAt: string;
}